requirements.txt
asgiref==3.5.2
asttokens==2.1.0
backcall==0.2.0
colorama==0.4.6
decorator==5.1.1
Django==4.1.3
django-cors-headers==3.13.0
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
executing==1.2.0
ipython==8.6.0
jedi==0.18.1
matplotlib-inline==0.1.6
parso==0.8.3
phonenumbers==8.13.0
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.32
pure-eval==0.2.2
Pygments==2.13.0
PyJWT==2.6.0
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.0
traitlets==5.5.0
tzdata==2022.6
wcwidth==0.2.5
